Add link to Dexcom app in G7 settings
I found myself missing the 'Open app' button from the G6, and that my muscle memory kept expecting it to be there, so in this PR I've added it into the G7 management screen.

I opted to create a new section rather than include it with the other management buttons beneath to try to prevent users accidentally tapping the more destructive actions.